Initial commit of my Simple Web Audio Player
This project has been in my personal Subversion repository
for over a decade but it is time to make it public.  Too
many people wanted to see the code and use it with their
own music collection.  (Especially some Tesla owners)

So, given that I have had a bit of time, I figured, why not
release it publicly.

The hardest part was making sure I did not have any "personal"
information in here and that it was not too specialized to my
environment.

The biggest change was to my Makefile as I had many local test
mechanisms that has to be removed from the public one as, well,
they assumed local web servers and ssh access and...

Still would like to get some of my benchmarking built into
here but that may be a bit of work yet.