Add support for making the Demo site
This way not only can I remake the Demo site, it
can be remade by others too, for their own use.

This works out as another way to rest but it also
adds the demo markers to the main html file.

Also cleaned up some CSS mess that I saw while
playing in the demo.